+++ packages/admin/resources/views/partials/orders/lines.blade.php
@@ -118,10 +118,27 @@
     </div>
 
     <div
-      class="pl-[calc(8rem_-_10px)] text-gray-700"
+      class="pl-24 text-gray-700"
       x-show="showDetails"
     >
-      <div class="pt-4 mt-4 space-y-4 border-t border-gray-200">
+      @if(!is_null($line->purchasable?->stock))
+      <span class="border rounded block p-2 text-xs mt-2">
+        <span @class([
+          'text-red-500' => $line->purchasable->stock < 50,
+          'text-green-500' => $line->purchasable->stock > 50
+        ])>
+          {{ __('adminhub::partials.orders.lines.current_stock_level', [
+            'count' => $line->purchasable->stock
+          ])}}
+        </span>
+        @if(!is_null($line->meta?->stock_level))
+          ({{ __('adminhub::partials.orders.lines.purchase_stock_level', [
+            'count' => $line->meta->stock_level
+          ])}})
+        @endif
+      </span>
+      @endif
+      <div class="mt-4 space-y-4">
         <article class="text-sm">
           <p>
             <strong>{{ __('adminhub::global.notes') }}:</strong>